Implement Predictive Demand Planning
- Added `app/demand_planner.py` with `FlightService` and `calculate_required_staff` logic.
- Updated `app/models.py` to include `demand_signal` in `OptimizeRequest` and changed `Shift.duration_hours` to float.
- Updated `app/optimizer.py` to generate shifts from `demand_signal`.
- Updated `app/main.py` with `/demand/{airport_code}` endpoint and integrated demand into optimizer.
- Added dependency `requests` to `requirements.txt`.

+++ app/models.py
@@ -25,7 +25,7 @@ class Shift(BaseModel):
     type: str  # "Morning", "Night"
     start_time: int
     end_time: int
-    duration_hours: int
+    duration_hours: float
     required_staff_count: int = 1
 
 class RosterAssignment(BaseModel):
